requirements.txt
Django==1.11.1
python_dateutil==2.7.0
djangorestframework==3.7.7
